Rename some names
