+++ textbot/__init__.py
@@ -1,7 +1,7 @@
 from .spellchecker import SpellChecker
 from .emailagent import EmailAgent
-from .nlp import tag
-from .emailaction import EmailAction
+from .router import tag
+from .action import EmailAction
 
 
 __version__ = "0.1.0"
